requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
avidtools==0.1.2
backoff==2.2.1
base2048==0.1.3
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cmd2==2.4.3
cohere==5.20.0
colorama==0.4.6
datasets==3.6.0
DateTime==5.5
deepl==1.17.0
diffusers==0.35.2
dill==0.3.8
distro==1.9.0
ecoji==0.1.1
eval_type_backport==0.3.0
fastapi==0.121.3
fastavro==1.12.1
fastuuid==0.14.0
filelock==3.20.0
frozenlist==1.8.0
fschat==0.2.36
fsspec==2025.3.0
ftfy==6.3.1
garak @ git+https://github.com/leondz/garak.git@fc829ff19ea60ce2c36b5160d079f25e92098dab
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-translate==3.23.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpath-python==1.1.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-core==0.3.80
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.46
latex2mathml==3.78.1
litellm==1.80.0
lorem==0.1.1
Markdown==3.10
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
mdurl==0.1.2
mistralai==1.5.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nemollm==0.3.5
networkx==3.6
nh3==0.3.2
nltk==3.9.2
numpy==2.3.5
nvdlib==0.8.3
nvidia-riva-client==2.16.0
ollama==0.6.1
openai==1.109.1
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
ply==3.11
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
replicate==1.0.7
requests==2.32.5
requests-futures==1.0.2
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
sentencepiece==0.2.1
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
stdlibs==2025.10.28
svgwrite==1.4.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.1
tqdm==4.67.1
transformers==4.56.2
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
wavedrom==2.0.3.post3
wcwidth==0.2.14
wn==0.9.5
xdg-base-dirs==6.0.2
xxhash==3.6.0
yarl==1.22.0
zalgolib==0.2.2
zipp==3.23.0
zope.interface==8.1.1
zstandard==0.25.0
